chore: refactor ESLint config
Consolidated ESLint configuration into eslint.config.mjs, removed the duplicate config file, and improved TypeScript linting setup. Added tsconfig.lint.json for linting purposes and updated tsconfig.json includes. Also updated VSCode settings for cSpell custom words.
