+++ .vscode/settings.json
@@ -1,1 +1,6 @@
-{}
+{
+  "cSpell.words": [
+    "styl",
+    "webp"
+  ]
+}
