+++ tsconfig.json
@@ -9,9 +9,7 @@
     "importHelpers": true,
     "sourceMap": true
   },
-  "include": [
-    "src"
-  ],
+  "include": [ "src" ],
   "files": [
     "src/resource.d.ts"
   ]
