+++ tsconfig.lint.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "target": "es2021",
+    "module": "commonjs",
+    "strict": true,
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+  },
+  "include": [ "*/**/*.ts", "*/**/*.js", "*/**/*.json", "webpack.config.js"],
+  "exclude": [ "node_modules", "dist" ]
+}
